Extract archive ID and data fetching from Archive.php to new ArchiveTableStore class.
