+++ core/Archive.php
@@ -8,10 +8,8 @@
  */
 namespace Piwik;
 
-use Piwik\Archive\IdArchiveCache;
+use Piwik\Archive\ArchiveTableStore;
 use Piwik\Archive\Parameters;
-use Piwik\ArchiveProcessor\Rules;
-use Piwik\Archive\ArchiveInvalidator;
 use Piwik\Container\StaticContainer;
 use Piwik\DataAccess\ArchiveSelector;
 use Piwik\Period\Factory as PeriodFactory;
@@ -135,19 +133,9 @@ class Archive
     private $params;
 
     /**
-     * @var \Piwik\Cache\Cache
+     * @var ArchiveTableStore
      */
-    private static $cache;
-
-    /**
-     * @var ArchiveInvalidator
-     */
-    private $invalidator;
-
-    /**
-     * @var IdArchiveCache
-     */
-    private $idArchiveCahe;
+    private $archiveTableStore;
 
     /**
      * @param Parameters $params
@@ -161,8 +149,7 @@ class Archive
         $this->forceIndexedBySite = $forceIndexedBySite;
         $this->forceIndexedByDate = $forceIndexedByDate;
 
-        $this->invalidator = StaticContainer::get(ArchiveInvalidator::class);
-        $this->idArchiveCahe = StaticContainer::get(IdArchiveCache::class);
+        $this->archiveTableStore = StaticContainer::get(ArchiveTableStore::class);
     }
 
     /**
@@ -385,23 +372,6 @@ class Archive
     }
 
     /**
-     * Returns the list of plugins that archive the given reports.
-     *
-     * @param array $archiveNames
-     * @return array
-     */
-    private function getRequestedPlugins($archiveNames)
-    {
-        $result = array();
-
-        foreach ($archiveNames as $name) {
-            $result[] = self::getPluginForReport($name);
-        }
-
-        return array_unique($result);
-    }
-
-    /**
      * Returns an object describing the set of sites, the set of periods and the segment
      * this Archive will query data for.
      *
@@ -460,68 +430,6 @@ class Archive
         return $dataTable;
     }
 
-    private function getSiteIdsThatAreRequestedInThisArchiveButWereNotInvalidatedYet()
-    {
-        if (is_null(self::$cache)) {
-            self::$cache = Cache::getTransientCache();
-        }
-
-        $id = 'Archive.SiteIdsOfRememberedReportsInvalidated';
-
-        if (!self::$cache->contains($id)) {
-            self::$cache->save($id, array());
-        }
-
-        $siteIdsAlreadyHandled = self::$cache->fetch($id);
-        $siteIdsRequested      = $this->params->getIdSites();
-
-        foreach ($siteIdsRequested as $index => $siteIdRequested) {
-            $siteIdRequested = (int) $siteIdRequested;
-
-            if (in_array($siteIdRequested, $siteIdsAlreadyHandled)) {
-                unset($siteIdsRequested[$index]); // was already handled previously, do not do it again
-            } else {
-                $siteIdsAlreadyHandled[] = $siteIdRequested; // we will handle this id this time
-            }
-        }
-
-        self::$cache->save($id, $siteIdsAlreadyHandled);
-
-        return $siteIdsRequested;
-    }
-
-    private function invalidatedReportsIfNeeded()
-    {
-        $siteIdsRequested = $this->getSiteIdsThatAreRequestedInThisArchiveButWereNotInvalidatedYet();
-
-        if (empty($siteIdsRequested)) {
-            return; // all requested site ids were already handled
-        }
-
-        $sitesPerDays = $this->invalidator->getRememberedArchivedReportsThatShouldBeInvalidated();
-
-        foreach ($sitesPerDays as $date => $siteIds) {
-            if (empty($siteIds)) {
-                continue;
-            }
-
-            $siteIdsToActuallyInvalidate = array_intersect($siteIds, $siteIdsRequested);
-
-            if (empty($siteIdsToActuallyInvalidate)) {
-                continue; // all site ids that should be handled are already handled
-            }
-
-            try {
-                $this->invalidator->markArchivesAsInvalidated($siteIdsToActuallyInvalidate, array(Date::factory($date)), false);
-            } catch (\Exception $e) {
-                Site::clearCache();
-                throw $e;
-            }
-        }
-
-        Site::clearCache();
-    }
-
     /**
      * Queries archive tables for data and returns the result.
      * @param array|string $archiveNames
@@ -553,13 +461,13 @@ class Archive
         $result = new Archive\DataCollection(
             $dataNames, $archiveDataType, $this->params->getIdSites(), $this->params->getPeriods(), $defaultRow = null);
 
-        $archiveIds = $this->getArchiveIds($archiveNames);
+        $archiveIds = $this->archiveTableStore->getArchiveIds($this->params, $archiveNames);
 
         if (empty($archiveIds)) {
             return $result;
         }
 
-        $archiveData = ArchiveSelector::getArchiveData($archiveIds, $archiveNames, $archiveDataType, $idSubtable);
+        $archiveData = $this->archiveTableStore->getArchiveData($archiveIds, $archiveNames, $archiveDataType, $idSubtable);
 
         $isNumeric = $archiveDataType == 'numeric';
 
@@ -579,141 +487,7 @@ class Archive
         return $result;
     }
 
-    /**
-     * Returns archive IDs for the sites, periods and archive names that are being
-     * queried. This function will use the idarchive cache if it has the right data,
-     * query archive tables for IDs w/o launching archiving, or launch archiving and
-     * get the idarchive from ArchiveProcessor instances.
-     *
-     * @param string[] $archiveNames
-     * @return array
-     */
-    private function getArchiveIds($archiveNames)
-    {
-        $plugins = $this->getRequestedPlugins($archiveNames);
 
-        // figure out which archives haven't been processed (if an archive has been processed,
-        // then we have the archive IDs in $this->idarchives)
-        $doneFlags     = array();
-        $archiveGroups = array();
-        foreach ($plugins as $plugin) {
-            $doneFlag = $this->getDoneStringForPlugin($plugin, $this->params->getIdSites());
-
-            $doneFlags[$doneFlag] = true;
-
-            $archiveGroup = $this->getArchiveGroupOfPlugin($plugin);
-
-            if ($archiveGroup == self::ARCHIVE_ALL_PLUGINS_FLAG) {
-                $archiveGroup = reset($plugins);
-            }
-            $archiveGroups[] = $archiveGroup;
-
-            $globalDoneFlag = Rules::getDoneFlagArchiveContainsAllPlugins($this->params->getSegment());
-            if ($globalDoneFlag !== $doneFlag) {
-                $doneFlags[$globalDoneFlag] = true;
-            }
-        }
-
-        $archiveGroups = array_unique($archiveGroups);
-
-        // cache id archives for plugins we haven't processed yet
-        if (!empty($archiveGroups)) {
-            if (!Rules::isArchivingDisabledFor($this->params->getIdSites(), $this->params->getSegment(), $this->getPeriodLabel())) {
-                $this->cacheArchiveIdsAfterLaunching($archiveGroups);
-            } else {
-                $this->cacheArchiveIdsWithoutLaunching($plugins);
-            }
-        }
-
-        $idArchivesByMonth = $this->getIdArchivesByMonth($doneFlags);
-
-        return $idArchivesByMonth;
-    }
-
-    /**
-     * Gets the IDs of the archives we're querying for and stores them in $this->archives.
-     * This function will launch the archiving process for each period/site/plugin if
-     * metrics/reports have not been calculated/archived already.
-     *
-     * @param array $archiveGroups @see getArchiveGroupOfReport
-     * @param array $plugins List of plugin names to archive.
-     */
-    private function cacheArchiveIdsAfterLaunching($archiveGroups)
-    {
-        $this->invalidatedReportsIfNeeded();
-
-        $today = Date::today();
-
-        foreach ($this->params->getPeriods() as $period) {
-            $twoDaysBeforePeriod = $period->getDateStart()->subDay(2);
-            $twoDaysAfterPeriod = $period->getDateEnd()->addDay(2);
-
-            foreach ($this->params->getIdSites() as $idSite) {
-                $site = new Site($idSite);
-
-                // if the END of the period is BEFORE the website creation date
-                // we already know there are no stats for this period
-                // we add one day to make sure we don't miss the day of the website creation
-                if ($twoDaysAfterPeriod->isEarlier($site->getCreationDate())) {
-                    Log::debug("Archive site %s, %s (%s) skipped, archive is before the website was created.",
-                        $idSite, $period->getLabel(), $period->getPrettyString());
-                    continue;
-                }
-
-                // if the starting date is in the future we know there is no visiidsite = ?t
-                if ($twoDaysBeforePeriod->isLater($today)) {
-                    Log::debug("Archive site %s, %s (%s) skipped, archive is after today.",
-                        $idSite, $period->getLabel(), $period->getPrettyString());
-                    continue;
-                }
-
-                $this->prepareArchive($archiveGroups, $site, $period);
-            }
-        }
-    }
-
-    /**
-     * Gets the IDs of the archives we're querying for and stores them in $this->archives.
-     * This function will not launch the archiving process (and is thus much, much faster
-     * than cacheArchiveIdsAfterLaunching).
-     *
-     * @param array $plugins List of plugin names from which data is being requested.
-     */
-    private function cacheArchiveIdsWithoutLaunching($plugins)
-    {
-        // TODO: if the archives already exist in the cache, we don't need to re-query
-        $idarchivesByReport = ArchiveSelector::getArchiveIds(
-            $this->params->getIdSites(), $this->params->getPeriods(), $this->params->getSegment(), $plugins);
-
-        foreach ($idarchivesByReport as $doneFlag => $idarchivesByDate) {
-            foreach ($idarchivesByDate as $dateRange => $idArchives) {
-                foreach ($idArchives as $idSite => $idArchive) {
-                    $this->idArchiveCahe->set($idSite, $dateRange, $doneFlag, $idArchive);
-                }
-            }
-        }
-    }
-
-    /**
-     * Returns the done string flag for a plugin using this instance's segment & periods.
-     * @param string $plugin
-     * @return string
-     */
-    private function getDoneStringForPlugin($plugin, $idSites)
-    {
-        return Rules::getDoneStringFlagFor(
-                    $idSites,
-                    $this->params->getSegment(),
-                    $this->getPeriodLabel(),
-                    $plugin
-        );
-    }
-
-    private function getPeriodLabel()
-    {
-        $periods = $this->params->getPeriods();
-        return reset($periods)->getLabel();
-    }
 
     /**
      * Returns an array describing what metadata to use when indexing a query result.
@@ -754,112 +528,5 @@ class Archive
         // Round up the value with 2 decimals
         // we cast the result as float because returns false when no visitors
         return round((float)$value, 2);
-    }
-
-    /**
-     * Returns the archiving group identifier given a plugin.
-     *
-     * More than one plugin can be called at once when archiving. In such a case
-     * we don't want to launch archiving three times for three plugins if doing
-     * it once is enough, so getArchiveIds makes sure to get the archive group of
-     * all reports.
-     *
-     * If the period isn't a range, then all plugins' archiving code is executed.
-     * If the period is a range, then archiving code is executed individually for
-     * each plugin.
-     */
-    private function getArchiveGroupOfPlugin($plugin)
-    {
-        $periods = $this->params->getPeriods();
-        $periodLabel = reset($periods)->getLabel();
-        
-        if (Rules::shouldProcessReportsAllPlugins($this->params->getIdSites(), $this->params->getSegment(), $periodLabel)) {
-            return self::ARCHIVE_ALL_PLUGINS_FLAG;
-        }
-
-        return $plugin;
-    }
-
-    /**
-     * Returns the name of the plugin that archives a given report.
-     *
-     * @param string $report Archive data name, eg, `'nb_visits'`, `'DevicesDetection_...'`, etc.
-     * @return string Plugin name.
-     * @throws \Exception If a plugin cannot be found or if the plugin for the report isn't
-     *                    activated.
-     */
-    private static function getPluginForReport($report)
-    {
-        // Core metrics are always processed in Core, for the requested date/period/segment
-        if (in_array($report, Metrics::getVisitsMetricNames())) {
-            $report = 'VisitsSummary_CoreMetrics';
-        } // Goal_* metrics are processed by the Goals plugin (HACK)
-        elseif (strpos($report, 'Goal_') === 0) {
-            $report = 'Goals_Metrics';
-        } elseif (strrpos($report, '_returning') === strlen($report) - strlen('_returning')) { // HACK
-            $report = 'VisitFrequency_Metrics';
-        }
-
-        $plugin = substr($report, 0, strpos($report, '_'));
-        if (empty($plugin)
-            || !\Piwik\Plugin\Manager::getInstance()->isPluginActivated($plugin)
-        ) {
-            throw new \Exception("Error: The report '$report' was requested but it is not available at this stage."
-                               . " (Plugin '$plugin' is not activated.)");
-        }
-        return $plugin;
-    }
-
-    /**
-     * @param $archiveGroups
-     * @param $site
-     * @param $period
-     */
-    private function prepareArchive(array $archiveGroups, Site $site, Period $period)
-    {
-        $parameters = new ArchiveProcessor\Parameters($site, $period, $this->params->getSegment());
-        $archiveLoader = new ArchiveProcessor\Loader($parameters);
-
-        $periodString = $period->getRangeString();
-
-        $idSite = $site->getId();
-        
-        // process for each plugin as well
-        foreach ($archiveGroups as $plugin) {
-            $doneFlag = $this->getDoneStringForPlugin($plugin, [$idSite]);
-            if ($this->idArchiveCahe->has($idSite, $periodString, $doneFlag)) {
-                continue;
-            }
-
-            $idArchive = $archiveLoader->prepareArchive($plugin);
-            $this->idArchiveCahe->set($idSite, $periodString, $doneFlag, $idArchive);
-        }
-    }
-
-    private function getIdArchivesByMonth($doneFlags)
-    {
-        // order idarchives by the table month they belong to
-        $idArchivesByMonth = array();
-
-        foreach (array_keys($doneFlags) as $doneFlag) {
-            foreach ($this->params->getPeriods() as $period) {
-                $dateRange = $period->getRangeString();
-                foreach ($this->params->getIdSites() as $idSite) {
-                    if ($this->idArchiveCahe->hasNonEmpty($idSite, $dateRange, $doneFlag)) {
-                        $idArchivesByMonth[$dateRange][] = $this->idArchiveCahe->get($idSite, $dateRange, $doneFlag);
-                    }
-                }
-            }
-        }
-
-        return $idArchivesByMonth;
-    }
-
-    /**
-     * @internal
-     */
-    public static function clearStaticCache()
-    {
-        self::$cache = null;
     }
 }
